requirements.txt
torch
transformers
numpy
matplotlib
jupyter
pyyaml